Fixes Clean up and moved Methods to other files
